Improve organizing of models
