+++ projects/fhi-angular-highcharts/src/lib/fhi-diagram.models.ts
@@ -1,11 +1,4 @@
 import { Options } from 'highcharts';
-import { FhiMapTypeId } from './fhi-diagram-type.constants';
-
-export interface TableHeaderCell {
-  name?: string;
-  colspan?: number;
-  rowspan?: number;
-}
 
 export interface FhiDiagramType {
   id: string;
@@ -13,53 +6,4 @@ export interface FhiDiagramType {
   name: string;
   group?: string;
   options?: Options;
-}
-
-export interface Data {
-  name: string;
-  y: number | string;
-}
-
-export interface FlagWithDataPointName extends FhiDiagramFlag {
-  name: string;
-}
-
-export interface FlaggedSerie {
-  name: string;
-  flaggedDataPoints: Array<FlagWithDataPointName>;
-}
-
-// Public interfaces
-
-export interface FhiDiagramFlag {
-  symbol: string;
-  label: string;
-}
-
-export interface FhiDiagramSerie {
-  data: Array<Data>;
-  name: string | Array<string>;
-  stack?: string;
-}
-
-export interface FhiDiagramOptions {
-  creditsHref?: string;
-  creditsText?: string;
-  diagramTypeId?: string;
-  diagramTypeNavId?: string;
-  diagramTypeSubset?: Array<string>;
-  disclaimer?: string;
-  flags?: Array<FhiDiagramFlag>;
-  lastUpdated?: string;
-  mapTypeId?: keyof typeof FhiMapTypeId;
-  openSource?: boolean;
-  series: Array<FhiDiagramSerie>;
-  title: string;
-}
-
-export interface FhiAllDiagramOptions extends FhiDiagramOptions {
-  allMapsLoaded?: boolean;
-  diagramType?: FhiDiagramType;
-  seriesHasDecimalDataPoints?: boolean;
-  seriesHasNegativeDataPoints?: boolean;
 }
