+++ projects/fhi-angular-highcharts/src/lib/models/all-diagram-options.models.ts
@@ -1,0 +1,9 @@
+import { FhiDiagramType } from '../fhi-diagram.models';
+import { FhiDiagramOptions } from './fhi-diagram-options.models';
+
+export interface AllDiagramOptions extends FhiDiagramOptions {
+  allMapsLoaded?: boolean;
+  diagramType?: FhiDiagramType;
+  seriesHasDecimalDataPoints?: boolean;
+  seriesHasNegativeDataPoints?: boolean;
+}
