+++ projects/fhi-angular-highcharts/src/lib/models/diagram-serie-data.models.ts
@@ -1,0 +1,4 @@
+export interface DiagramSerieData {
+  name: string;
+  y: number | string;
+}
