+++ projects/fhi-angular-highcharts/src/lib/models/fhi-diagram-flag.models.ts
@@ -1,0 +1,4 @@
+export interface FhiDiagramFlag {
+  symbol: string;
+  label: string;
+}
