+++ projects/fhi-angular-highcharts/src/lib/models/fhi-diagram-serie.models.ts
@@ -1,0 +1,7 @@
+import { DiagramSerieData } from './diagram-serie-data.models';
+
+export interface FhiDiagramSerie {
+  data: Array<DiagramSerieData>;
+  name: string | Array<string>;
+  stack?: string;
+}
