+++ projects/fhi-angular-highcharts/src/lib/models/flag-with-data-point-name.models.ts
@@ -1,0 +1,5 @@
+import { FhiDiagramFlag } from './fhi-diagram-flag.models';
+
+export interface FlagWithDataPointName extends FhiDiagramFlag {
+  name: string;
+}
