+++ projects/fhi-angular-highcharts/src/lib/models/flagged-serie.models.ts
@@ -1,0 +1,6 @@
+import { FlagWithDataPointName } from './flag-With-data-point-name.models';
+
+export interface FlaggedSerie {
+  name: string;
+  flaggedDataPoints: Array<FlagWithDataPointName>;
+}
