+++ projects/fhi-angular-highcharts/src/public-api.ts
@@ -5,4 +5,6 @@
 export * from './lib/fhi-angular-highcharts.component';
 export * from './lib/fhi-angular-highcharts.module';
 
-export { FhiDiagramOptions, FhiDiagramFlag, FhiDiagramSerie } from './lib/fhi-diagram.models';
+export * from './lib/models/fhi-diagram-flag.models';
+export * from './lib/models/fhi-diagram-options.models';
+export * from './lib/models/fhi-diagram-serie.models';
